General review update
 Thanks for putting everything together. I have reviewed and agree that the list covers everything we need for the PR-D/X submission.

I have just added "unit" tests for clarification and reference of Dr. Ravi and whoever he decides to assign. I also have the following propositions:

1) It would be best to create a sequence of prioritization for all the tasks at hand 
2) It would also be easy to coordinate in the best way if we do a task assignment and distribution of ownership to address everything in this limited time-frame. I can set up a Kanban-board with the entire task list and then parts of it could be taken up or assigned to each of us in a specific sequence for clarity, efficiency and coordination. 

This would be the fastest way to coordinate , distribute responsibilities and take things forward in my opinion. I could set-up an initial Kanban on Notion and share the link for everyone's reference. We could move it to Github later if that is more suitable for everyone.